LPD-98209 Rename dashboard SpacesDropdown to SpacePicker
