+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/dashboard/performance/components/Filters.tsx
@@ -10,7 +10,7 @@ import {
 } from '@liferay/analytics-reports-js-components-web';
 import React, {useContext} from 'react';
 
-import {SpacesDropdown} from '../../common/SpacesDropdown';
+import {SpacePicker} from '../../common/SpacePicker';
 import {PerformanceContext} from '../PerformanceContext';
 
 export function Filters() {
@@ -20,7 +20,7 @@ export function Filters() {
 		<ClayLayout.Row className="mb-4">
 			<ClayLayout.Col size={12}>
 				<div className="d-flex">
-					<SpacesDropdown
+					<SpacePicker
 						className="mr-3"
 						onSelectSpace={setSpace}
 						selectedSpace={space}
