+++ modules/apps/site/site-cms-site-initializer/test/js/main_view/dashboard/LanguagesDropdown.test.tsx
@@ -14,7 +14,7 @@ import {
 import React from 'react';
 
 import SpaceService from '../../../../src/main/resources/META-INF/resources/js/common/services/SpaceService';
-import {initialSpace} from '../../../../src/main/resources/META-INF/resources/js/main_view/dashboard/common/SpacesDropdown';
+import {initialSpace} from '../../../../src/main/resources/META-INF/resources/js/main_view/dashboard/common/SpacePicker';
 import {
 	InventoryContext,
 	initialLanguage,
